NO-ISSUE: test: remove shared/ directory, move utils to canonical locations
csrf.ts moves to utils/api/ (its only consumers) and test-utils.ts moves
to utils/. The shared/ directory existed to bridge two separate test suites
which are now unified.
